src/user_service_v1/requirements.txt
python-dotenv==1.0.1
Flask==3.1.0
flask-restx==1.3.0
pymongo==4.10.1
gunicorn==23.0.0
pika==1.3.2
